backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.6
openai>=1.0.0
requests>=2.31.0
python-chess==1.999
tenacity>=8.2.3
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
